feat: better parsing numeric types (builtin function)

+++ WitLang-Model/src/main/kotlin/io/github/slupik/schemablock/newparser/function/definition/FunctionAsLong.java
@@ -35,7 +35,7 @@ public class FunctionAsLong implements Function {
         try {
             return new SimpleValueImpl(
                     ValueType.LONG,
-                    Long.parseLong(value.toString())
+                    ((long) Double.parseDouble(value.toString()))
             );
         } catch (Exception e) {
             throw new CannotParseData(value, ValueType.LONG, line, position);
